requirements-updated.txt
torch==1.13.1
torchvision==0.14.1
diffusers==0.14.0
transformers==4.27.1
accelerate==0.18.0
qrcode==7.4.2
Pillow==9.4.0
numpy==1.24.2
gradio==3.23.0
